Update ExpertPhotom_GWCedits_15mar24.py
Updated LogNorm to be SymLogNorm in almost all of the figures because they contain very small negative values.

This fixes the images that are too pale or otherwise experience problems being graphed while also making darker images easier to see. 

The one exception is the imshow that used make_lupton_rgb as that function does not need a LogNorm and crashes if it uses one. 

LinearThreshold is then the constant value for what degree of the function is linear; since the negative values in the images are very small, a small linear threshold works well. It is virtually indistinguishable from 1-0.00001 and likely beyond, so the actual value does not matter much.