Update make_plots.py
Combined KDE plotting and Histogram plotting code into one to clean up appearance, removed hardcoding on plots to be more brief and easy to read. Several formatting fixes and combination of CORNISH N and CORNISH S data, as well as making formatting between histograms and KDE plots consistent and easy to read. The code should be easily accessible so that all you have to do is adjust the csv file name to have your ID to get it to generate plots from your data.